# src/main/java/com/ecommerce/nashtech/dto/order/OrderItemResponse.java
package com.ecommerce.nashtech.dto.order;

import com.ecommerce.nashtech.dto.perfume.PerfumeResponse;
import lombok.Data;

@Data
public class OrderItemResponse {
    private Long id;
    private Long amount;
    private Long quantity;
    private PerfumeResponse perfume;
}