# src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "HaoHaoShop"
servers:
  - url: "http://localhost:8080"
...
paths:
  /api/v1/admin/add:
    post:
      tags:
        - "AdminController"
      summary: "Add perfume to database"
      operationId: "addPerfume"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                perfumeTitle:
                  type: string
                perfumer:
                  type: string
                year:
                  type: string
                country:
                  type: string
                perfumeGender:
                  type: string
                fragranceTopNotes:
                  type: string
                fragranceMiddleNotes:
                  type: string
                fragranceBaseNotes:
                  type: string
                description:
                  type: string
                filename:
                  type: string
                price:
                  type: string
                volume:
                  type: string
                type:
                  type: string
              required:
                - perfumeTitle
                - perfumer
                - year
                - country
                - perfumeGender
                - fragranceTopNotes
                - fragranceMiddleNotes
                - fragranceBaseNotes
                - price
                - volume
                - type
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeResponse"
              examples:
                PerfumeResponse:
                  $ref: "#/components/schemas/PerfumeResponse"
        "400":
          description: "Bad request"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeError"
              examples:
                PerfumeError:
                  $ref: "#/components/schemas/PerfumeError"
      security:
        - token: []
  /api/v1/admin/edit:
    put:
      tags:
        - "AdminController"
      summary: "Update perfume"
      operationId: "updatePerfume"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                perfumeTitle:
                  type: string
                perfumer:
                  type: string
                year:
                  type: string
                country:
                  type: string
                perfumeGender:
                  type: string
                fragranceTopNotes:
                  type: string
                fragranceMiddleNotes:
                  type: string
                fragranceBaseNotes:
                  type: string
                description:
                  type: string
                filename:
                  type: string
                price:
                  type: string
                volume:
                  type: string
                type:
                  type: string
              required:
                - perfumeTitle
                - perfumer
                - year
                - country
                - perfumeGender
                - fragranceTopNotes
                - fragranceMiddleNotes
                - fragranceBaseNotes
                - price
                - volume
                - type
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeResponse"
              examples:
                PerfumeResponse:
                  $ref: "#/components/schemas/PerfumeResponse"
        "400":
          description: "Bad request"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeError"
              examples:
                PerfumeError:
                  $ref: "#/components/schemas/PerfumeError"
      security:
        - token: []
  /api/v1/admin/delete/{id}:
    delete:
      tags:
        - "AdminController"
      summary: "Delete perfume"
      operationId: "deletePerfume"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
      security:
        - token: []
  /api/v1/admin/orders:
    get:
      tags:
        - "AdminController"
      summary: "Get all orders from database"
      operationId: "getAllOrders"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/order:
    post:
      tags:
        - "AdminController"
      summary: "Get user order by email"
      operationId: "getUserOrdersByEmail"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test123@test.com"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/order/delete/{id}:
    delete:
      tags:
        - "AdminController"
      summary: "Delete order"
      operationId: "deleteOrder"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/user/{id}:
    get:
      tags:
        - "AdminController"
      summary: "Get user by id"
      operationId: "getUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
      security:
        - token: []
  /api/v1/admin/user/all:
    get:
      tags:
        - "AdminController"
      summary: "Get users from database"
      operationId: "getAllUsers"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
      security:
        - token: []
  /api/v1/admin/graphql/user:
    post:
      tags:
        - "AdminController"
      summary: "Get user by GraphQL query"
      operationId: "getUserByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ user(id: 2) { id email firstName lastName city address phoneNumber postIndex } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  email:
                    type: "string"
                  firstName:
                    type: "string"
                  lastName:
                    type: "string"
                  city:
                    type: "string"
                  address:
                    type: "string"
                  phoneNumber:
                    type: "string"
                  postIndex:
                    type: "string"
                example:
                  id: 2
                  email: "test123@test.com"
                  firstName: "John"
                  lastName: "Doe"
                  city: "New York"
                  address: "Wall Street 1"
                  phoneNumber: "1234567890"
                  postIndex: "1234567890"
      security:
        - token: []
  /api/v1/admin/graphql/user/all:
    post:
      tags:
        - "AdminController"
      summary: "Get all users by GraphQL query"
      operationId: "getAllUsersByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ users { id email firstName lastName city address phoneNumber postIndex } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    email:
                      type: "string"
                    firstName:
                      type: "string"
                    lastName:
                      type: "string"
                    city:
                      type: "string"
                    address:
                      type: "string"
                    phoneNumber:
                      type: "string"
                    postIndex:
                      type: "string"
                  example:
                    id: 2
                    email: "test123@test.com"
                    firstName: "John"
                    lastName: "Doe"
                    city: "New York"
                    address: "Wall Street 1"
                    phoneNumber: "1234567890"
                    postIndex: "1234567890"
      security:
        - token: []
  /api/v1/admin/graphql/orders:
    post:
      tags:
        - "AdminController"
      summary: "Get all orders by GraphQL query"
      operationId: "getAllOrdersQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ orders { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/graphql/order:
    post:
      tags:
        - "AdminController"
      summary: "Get orders by user Email GraphQL query"
      operationId: "getUserOrdersByEmailQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    PerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        year:
          type: "integer"
          format: "int32"
          example: 2015
        country:
          type: "string"
          example: "France"
        perfumeGender:
          type: "string"
          example: "male"
        fragranceTopNotes:
          type: "string"
          example: "Sichuan Pepper"
        fragranceMiddleNotes:
          type: "string"
          example: "Bergamot"
        fragranceBaseNotes:
          type: "string"
          example: "Ambroxan"
        description:
          type: "string"
          example: "Perfume description"
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        price:
          type: "integer"
          format: "int32"
          example: 62
        volume:
          type: "string"
          example: "100"
        type:
          type: "string"
          example: "Eau de parfum"
        reviews:
          type: "array"
          items:
            $ref: "#/components/schemas/ReviewResponse"
        file:
          type: "string"
          format: binary
    ReviewResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        author:
          type: "string"
          example: "John Doe"
        message:
          type: "string"
          example: "Message"
        date:
          type: "string"
          example: "2021-04-07"
    PerfumeError:
      type: "object"
      properties:
        perfumeTitleError:
          type: "string"
          example: "Fill in the input field"
        perfumerError:
          type: "string"
          example: "Fill in the input field"
        yearError:
          type: "string"
          example: "Fill in the input field"
        countryError:
          type: "string"
          example: "Fill in the input field"
        typeError:
          type: "string"
          example: "Fill in the input field"
        volumeError:
          type: "string"
          example: "Fill in the input field"
        perfumeGenderError:
          type: "string"
          example: "Fill in the input field"
        fragranceTopNotesError:
          type: "string"
          example: "Fill in the input field"
        fragranceMiddleNotesError:
          type: "string"
          example: "Fill in the input field"
        fragranceBaseNotesError:
          type: "string"
          example: "Fill in the input field"
        priceError:
          type: "string"
          example: "Fill in the input field"
    OrderResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        totalPrice:
          type: "number"
          format: "double"
          example: 100
        date:
          type: "string"
          example: "2021-04-07"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        email:
          type: "string"
          example: "test123@test.com"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "integer"
          format: "int32"
          example: 123456
        orderItems:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderItemResponse"
    OrderItemResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        amount:
          type: "integer"
          format: "int64"
          example: 100
        quantity:
          type: "integer"
          format: "int64"
          example: 2
        perfume:
          $ref: "#/components/schemas/PerfumeResponse"
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 2
        email:
          type: "string"
          example: "test123@test.com"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "string"
          example: "123456"
        provider:
          type: "string"
          example: "LOCAL"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
...